wificond: Add Gbk Ssid aging mechanism
Wifigbk maintains a list of <gbkSsid, utfSsid> pair in conversion
history, and it is never cleared unless device reboots.

This works if AP only uses Chinese Ssid with Gbk encoding, and
wifigbk converts the Ssid in connect request to Gbk as expected.
But when that AP is powered off and another AP is setup with exact
same Chinese SSID with UTF-8 encoding, wifigbk still converts the
Ssid to Gbk due to <gbkSsid, utfSsid> pair not removed from
conversion history.

Add Gbk Ssid aging mechanism, which removes the <gbkSsid, utfSsid>
pair from conversion history if gbkSsid not present in scan results.

Change-Id: Ic9e97fbde14b8e17876897ffdd3e984bf7b163ec
CRs-Fixed: 2184021